Fix bug in the calcluation of totalRelevantRecords
This bug was causing recall values to be much lower than expected
and was caused by duplicated values of Dataset_ID being present
in the data.frame 'filtered_merged_result' prior to the calculation
of totalRelevantRecords. This change filters filtered_merged_result
prior to the calculation to just the unique combinations of
Dataset_ID and whatever query we're looking at (q1, q2, etc).